Discovered wx was disabling a number of warnings. Cleanup

+++ src/Win/DlgConfigEventMethod.cpp
@@ -105,8 +105,8 @@ int wxCALLBACK ComparePlacement(wxIntPtr item1, wxIntPtr item2, wxIntPtr sortDat
 int wxCALLBACK ComparePlacement(long item1, long item2, long sortData)
 #endif
 {
-	CDlgConfigureDataPlacementPtr place1 = s_SortInfo.pThis->GetPlacementDataByData(item1);
-	CDlgConfigureDataPlacementPtr place2 = s_SortInfo.pThis->GetPlacementDataByData(item2);
+	CDlgConfigureDataPlacementPtr place1 = s_SortInfo.pThis->GetPlacementDataByData(static_cast<long>(item1));
+	CDlgConfigureDataPlacementPtr place2 = s_SortInfo.pThis->GetPlacementDataByData(static_cast<long>(item2));
 	return place1->Place() > place2->Place();
 }
